# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CtrlTrade SaaS Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/integrations/{provider}/connect:
    post:
      operationId: connectIntegration
      tags: [integrations]
      summary: Start the OAuth handshake for the given provider.
      parameters:
        - in: path
          name: provider
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [authUrl]
                properties:
                  authUrl: { type: string }
  /v1/integrations/{provider}:
    delete:
      operationId: disconnectIntegration
      tags: [integrations]
      summary: Disconnect the integration (tokens are revoked locally).
      parameters:
        - in: path
          name: provider
          required: true
          schema: { type: string }
      responses:
        "204": { description: No content }
  /v1/integrations/{provider}/sync:
    post:
      operationId: triggerIntegrationSync
      tags: [integrations]
      summary: Queue a manual sync for the given provider.
      parameters:
        - in: path
          name: provider
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [enqueued]
                properties:
                  enqueued: { type: boolean }
  /v1/integrations/{provider}/logs:
    get:
      operationId: getIntegrationLogs
      tags: [integrations]
      summary: Recent sync log entries for a provider.
      parameters:
        - in: path
          name: provider
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/IntegrationSyncLog" }
...
  /v1/admin/integration-catalogue/{provider}:
    put:
      operationId: updateAdminIntegrationCatalogue
      tags: [admin, integrations]
      summary: Enable, disable, or restrict a provider in the catalogue.
      parameters:
        - in: path
          name: provider
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminIntegrationCatalogueInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [provider, enabled]
                properties:
                  provider: { type: string }
                  enabled: { type: boolean }
                  minPlan: { type: ["string", "null"] }
                  updatedAt: { type: string, format: date-time }
...
components:
  schemas:
...
    IntegrationSyncLog:
      type: object
      required: [id, direction, status, createdAt]
      properties:
        id: { type: string }
        direction: { type: string }
        entityKind: { type: ["string", "null"] }
        entityId: { type: ["string", "null"] }
        status: { type: string }
        message: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
...
    AdminIntegrationCatalogueInput:
      type: object
      required: [enabled]
      properties:
        enabled: { type: boolean }
        minPlan: { type: ["string", "null"] }