# lib/api-zod/src/generated/types/adminIntegrationCatalogueInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CtrlTrade SaaS Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminIntegrationCatalogueInput {
  enabled: boolean;
  /** @nullable */
  minPlan?: string | null;
}