# lib/api-zod/src/generated/types/integrationSyncLog.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CtrlTrade SaaS Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface IntegrationSyncLog {
  id: string;
  direction: string;
  /** @nullable */
  entityKind?: string | null;
  /** @nullable */
  entityId?: string | null;
  status: string;
  /** @nullable */
  message?: string | null;
  createdAt: Date;
}